refactored endpoint communication code
