+++ src/SFA.DAS.Payments.Application/Infrastructure/Messaging/EndpointCommunicationSender.cs
@@ -3,7 +3,7 @@ using Autofac;
 using NServiceBus;
 using SFA.DAS.Payments.Messages.Core;
 
-namespace SFA.DAS.Payments.ServiceFabric.Core
+namespace SFA.DAS.Payments.Application.Infrastructure.Messaging
 {
     public class EndpointCommunicationSender : EndpointCommunicationBase, IEndpointCommunicationSender
     {
