+++ src/SFA.DAS.Payments.Application/Infrastructure/Messaging/IEndpointCommunicationSender.cs
@@ -1,7 +1,7 @@
 ﻿using System.Threading.Tasks;
 using SFA.DAS.Payments.Messages.Core;
 
-namespace SFA.DAS.Payments.ServiceFabric.Core
+namespace SFA.DAS.Payments.Application.Infrastructure.Messaging
 {
     public interface IEndpointCommunicationSender
     {
