--- src/SFA.DAS.Payments.RequiredPayments.EmployerRequiredPaymentsService.Interfaces/SFA.DAS.Payments.RequiredPayments.EmployerRequiredPaymentsService.Interfaces.csproj
@@ -1,14 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
-    <RuntimeFrameworkVersion>2.1.2</RuntimeFrameworkVersion>
-    <IsServiceFabricServiceProject>True</IsServiceFabricServiceProject>
-    <RuntimeIdentifier>win7-x64</RuntimeIdentifier>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.ServiceFabric.Actors" Version="3.2.162" />
-  </ItemGroup>
-
-</Project>
