--- src/SFA.DAS.Payments.RequiredPayments.EmployerRequiredPaymentsService/ActorEventSource.cs
@@ -1,169 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Diagnostics.Tracing;
-using System.Fabric;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Microsoft.ServiceFabric.Actors.Runtime;
-
-namespace SFA.DAS.Payments.RequiredPayments.EmployerRequiredPaymentsService
-{
-    [EventSource(Name = "MyCompany-SFA.DAS.Payments.RequiredPayments.ServiceFabric-SFA.DAS.Payments.RequiredPayments.EmployerRequiredPaymentsService")]
-    internal sealed class ActorEventSource : EventSource
-    {
-        public static readonly ActorEventSource Current = new ActorEventSource();
-
-        static ActorEventSource()
-        {
-            // A workaround for the problem where ETW activities do not get tracked until Tasks infrastructure is initialized.
-            // This problem will be fixed in .NET Framework 4.6.2.
-            Task.Run(() => { });
-        }
-
-        // Instance constructor is private to enforce singleton semantics
-        private ActorEventSource() : base() { }
-
-        #region Keywords
-        // Event keywords can be used to categorize events. 
-        // Each keyword is a bit flag. A single event can be associated with multiple keywords (via EventAttribute.Keywords property).
-        // Keywords must be defined as a public class named 'Keywords' inside EventSource that uses them.
-        public static class Keywords
-        {
-            public const EventKeywords HostInitialization = (EventKeywords)0x1L;
-        }
-        #endregion
-
-        #region Events
-        // Define an instance method for each event you want to record and apply an [Event] attribute to it.
-        // The method name is the name of the event.
-        // Pass any parameters you want to record with the event (only primitive integer types, DateTime, Guid & string are allowed).
-        // Each event method implementation should check whether the event source is enabled, and if it is, call WriteEvent() method to raise the event.
-        // The number and types of arguments passed to every event method must exactly match what is passed to WriteEvent().
-        // Put [NonEvent] attribute on all methods that do not define an event.
-        // For more information see https://msdn.microsoft.com/en-us/library/system.diagnostics.tracing.eventsource.aspx
-
-        [NonEvent]
-        public void Message(string message, params object[] args)
-        {
-            if (this.IsEnabled())
-            {
-                string finalMessage = string.Format(message, args);
-                Message(finalMessage);
-            }
-        }
-
-        private const int MessageEventId = 1;
-        [Event(MessageEventId, Level = EventLevel.Informational, Message = "{0}")]
-        public void Message(string message)
-        {
-            if (this.IsEnabled())
-            {
-                WriteEvent(MessageEventId, message);
-            }
-        }
-
-        [NonEvent]
-        public void ActorMessage(Actor actor, string message, params object[] args)
-        {
-            if (this.IsEnabled()
-                && actor.Id != null
-                && actor.ActorService != null
-                && actor.ActorService.Context != null
-                && actor.ActorService.Context.CodePackageActivationContext != null)
-            {
-                string finalMessage = string.Format(message, args);
-                ActorMessage(
-                    actor.GetType().ToString(),
-                    actor.Id.ToString(),
-                    actor.ActorService.Context.CodePackageActivationContext.ApplicationTypeName,
-                    actor.ActorService.Context.CodePackageActivationContext.ApplicationName,
-                    actor.ActorService.Context.ServiceTypeName,
-                    actor.ActorService.Context.ServiceName.ToString(),
-                    actor.ActorService.Context.PartitionId,
-                    actor.ActorService.Context.ReplicaId,
-                    actor.ActorService.Context.NodeContext.NodeName,
-                    finalMessage);
-            }
-        }
-
-        // For very high-frequency events it might be advantageous to raise events using WriteEventCore API.
-        // This results in more efficient parameter handling, but requires explicit allocation of EventData structure and unsafe code.
-        // To enable this code path, define UNSAFE conditional compilation symbol and turn on unsafe code support in project properties.
-        private const int ActorMessageEventId = 2;
-        [Event(ActorMessageEventId, Level = EventLevel.Informational, Message = "{9}")]
-        private
-    #if UNSAFE
-            unsafe
-    #endif
-            void ActorMessage(
-            string actorType,
-            string actorId,
-            string applicationTypeName,
-            string applicationName,
-            string serviceTypeName,
-            string serviceName,
-            Guid partitionId,
-            long replicaOrInstanceId,
-            string nodeName,
-            string message)
-        {
-    #if !UNSAFE
-            WriteEvent(
-                    ActorMessageEventId,
-                    actorType,
-                    actorId,
-                    applicationTypeName,
-                    applicationName,
-                    serviceTypeName,
-                    serviceName,
-                    partitionId,
-                    replicaOrInstanceId,
-                    nodeName,
-                    message);
-    #else
-                const int numArgs = 10;
-                fixed (char* pActorType = actorType, pActorId = actorId, pApplicationTypeName = applicationTypeName, pApplicationName = applicationName, pServiceTypeName = serviceTypeName, pServiceName = serviceName, pNodeName = nodeName, pMessage = message)
-                {
-                    EventData* eventData = stackalloc EventData[numArgs];
-                    eventData[0] = new EventData { DataPointer = (IntPtr) pActorType, Size = SizeInBytes(actorType) };
-                    eventData[1] = new EventData { DataPointer = (IntPtr) pActorId, Size = SizeInBytes(actorId) };
-                    eventData[2] = new EventData { DataPointer = (IntPtr) pApplicationTypeName, Size = SizeInBytes(applicationTypeName) };
-                    eventData[3] = new EventData { DataPointer = (IntPtr) pApplicationName, Size = SizeInBytes(applicationName) };
-                    eventData[4] = new EventData { DataPointer = (IntPtr) pServiceTypeName, Size = SizeInBytes(serviceTypeName) };
-                    eventData[5] = new EventData { DataPointer = (IntPtr) pServiceName, Size = SizeInBytes(serviceName) };
-                    eventData[6] = new EventData { DataPointer = (IntPtr) (&partitionId), Size = sizeof(Guid) };
-                    eventData[7] = new EventData { DataPointer = (IntPtr) (&replicaOrInstanceId), Size = sizeof(long) };
-                    eventData[8] = new EventData { DataPointer = (IntPtr) pNodeName, Size = SizeInBytes(nodeName) };
-                    eventData[9] = new EventData { DataPointer = (IntPtr) pMessage, Size = SizeInBytes(message) };
-
-                    WriteEventCore(ActorMessageEventId, numArgs, eventData);
-                }
-    #endif
-        }
-
-        private const int ActorHostInitializationFailedEventId = 3;
-        [Event(ActorHostInitializationFailedEventId, Level = EventLevel.Error, Message = "Actor host initialization failed", Keywords = Keywords.HostInitialization)]
-        public void ActorHostInitializationFailed(string exception)
-        {
-            WriteEvent(ActorHostInitializationFailedEventId, exception);
-        }
-        #endregion
-
-        #region Private Methods
-    #if UNSAFE
-            private int SizeInBytes(string s)
-            {
-                if (s == null)
-                {
-                    return 0;
-                }
-                else
-                {
-                    return (s.Length + 1) * sizeof(char);
-                }
-            }
-    #endif
-        #endregion
-    }
-}
