--- src/SFA.DAS.Payments.RequiredPayments.EmployerRequiredPaymentsService/PackageRoot/Config/Settings.xml
@@ -1,11 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Settings xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns="http://schemas.microsoft.com/2011/01/fabric">
-  <Section Name="EmployerRequiredPaymentsServiceActorServiceReplicatorConfig">
-    <Parameter Name="ReplicatorEndpoint" Value="EmployerRequiredPaymentsServiceActorServiceReplicatorEndpoint" />
-    <Parameter Name="BatchAcknowledgementInterval" Value="0.005" />
-  </Section>
-  <Section Name="EmployerRequiredPaymentsServiceActorServiceReplicatorSecurityConfig">
-    <Parameter Name="CredentialType" Value="None" />
-  </Section>
-  <!-- The content will be generated during build -->
-</Settings>
