+++ src/SFA.DAS.Payments.RequiredPayments.ServiceFabric/ApplicationParameters/Local.1Node.xml
@@ -6,5 +6,6 @@
     <Parameter Name="RequiredPaymentsServiceActorService_PartitionCount" Value="1" />
     <Parameter Name="RequiredPaymentsServiceActorService_TargetReplicaSetSize" Value="1" />
     <Parameter Name="RequiredPaymentsServiceActorService_MinReplicaSetSize" Value="1" />
+    <Parameter Name="StorageConnectionString" Value="UseDevelopmentStorage=true" />
   </Parameters>
 </Application>
