+++ src/SFA.DAS.Payments.RequiredPayments.ServiceFabric/ApplicationParameters/Local.5Node.xml
@@ -4,5 +4,7 @@
     <Parameter Name="SFA.DAS.Payments.RequiredPayments.RequiredPaymentsProxyService_InstanceCount" Value="1" />
     <Parameter Name="ProviderPaymentsService_InstanceCount" Value="-1" />
     <Parameter Name="RequiredPaymentsServiceActorService_PartitionCount" Value="1" />
+    <Parameter Name="StorageConnectionString" Value="UseDevelopmentStorage=true" />
+    <Parameter Name="LoggingConnectionString" Value="Server=.;Database=AppLog;User Id=SFActor; Password=SFActor;" />
   </Parameters>
 </Application>
