+++ src/SFA.DAS.Payments.Service.Core/IEndpointCommunicationListener.cs
@@ -2,7 +2,7 @@
 
 namespace SFA.DAS.Payments.ServiceFabric.Core
 {
-    public interface IEndpointCommunicationListener<T> : ICommunicationListener
+    public interface IEndpointCommunicationListener : ICommunicationListener
     {
     }
 }
