+++ src/SFA.DAS.Payments.Service.Core/SFA.DAS.Payments.ServiceFabric.Core.csproj
@@ -1,8 +1,14 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>netstandard2.0</TargetFramework>
   </PropertyGroup>
+
+  <ItemGroup>
+    <Compile Remove="EndpointCommunicationBase.cs" />
+    <Compile Remove="EndpointCommunicationSender.cs" />
+    <Compile Remove="IEndpointCommunicationSender.cs" />
+  </ItemGroup>
 
   <ItemGroup>
     <PackageReference Include="Autofac" Version="4.8.1" />
